3D Menu
I made this 3D menu using only HTML and CSS, making use of the Isometric Menu Hover Effects, to give that 3D appearance and movement to the side menu bar. It can be used as a stylization of a web page to give more dynamics and provide an interactive interface for the user, bringing a more modern appearance and a dynamic effect.